feat(search): added confing.class for Search, changed a logical of usage this settings, second try to solve the issue with search logical creation

+++ src/main/java/searchengine/model/DBIndex.java
@@ -32,5 +32,5 @@ public class DBIndex {
     private DBLemma dbLemma;
 
     @Column(name = "lemma_rank", nullable = false)
-    private float rank;
+    private int rank;
 }
